Refactor MyGroups with GroupCard
