+++ packages/lib-user/src/components/MyGroups/components/GroupCard/GroupCard.mock.js
