+++ packages/lib-user/src/components/MyGroups/components/GroupCard/GroupCard.stories.js
